+++ src/delete_vertex.c
@@ -6,7 +6,7 @@
 /*   By: sipatry <sipatry@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/20 17:10:35 by sipatry           #+#    #+#             */
-/*   Updated: 2019/11/20 17:10:39 by sipatry          ###   ########.fr       */
+/*   Updated: 2020/02/25 14:23:23 by lnicosia         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,8 @@ int		modify_sectors(t_env *env, int vertex)
 		{
 			if (env->sectors[i].vertices[j] == vertex)
 			{
-				env->sectors[i].vertices = ft_delindex(env->sectors[i].vertices,
+				env->sectors[i].vertices = (int*)ft_delindex(
+				env->sectors[i].vertices,
 					sizeof(int) * (env->sectors[i].nb_vertices + 1),
 					sizeof(int),
 					sizeof(int) * j);
@@ -50,22 +51,26 @@ int		delete_vertex(void *param)
 
 	env = (t_env*)param;
 	vertex = env->editor.selected_vertex;
-	env->vertices = ft_delindex(env->vertices,
+	env->vertices = (t_vertex*)ft_delindex(env->vertices,
 			sizeof(t_vertex) * env->nb_vertices,
 			sizeof(t_vertex),
 			sizeof(t_vertex) * vertex);
 	env->nb_vertices--;
+	if (env->nb_vertices > 0 && !env->vertices)
+		return (-1);
 	i = vertex;
 	while (i < env->nb_vertices)
 	{
 		env->vertices[i].num--;
 		i++;
 	}
-	if (env->nb_sectors)
-		if (modify_sectors(env, vertex))
-			return (-1);
 	if (delete_invalid_sectors(env))
 		return (-1);
+	if (env->nb_sectors)
+	{
+		if (modify_sectors(env, vertex))
+			return (-1);
+	}
 	if (delete_invalid_vertices(env))
 		return (-1);
 	clear_portals(env);
